chore: drop wee_alloc for WASM target
wee_alloc is unmaintained and has a lingering memory leak issue. There
are other alternative allocators, but the one provided by the Rust
standard library balances performance, being maintained and reviewed and
code size pretty well, so use it. This should have a cost of a few KiB
in the WASM executable size, but provides some peace of mind.