perf(wasm): switch to `rlsf` allocator
Commit 6156498d040a4e088076ec262669a4814deb7c2d dropped `wee_alloc` in
favor of the default memory allocator for WebAssembly targets,
`dlmalloc`. While this was warranted due to the bad maintenance status
of `wee_alloc`, which only got worse lately due to the sunsetting of the
rustwasm organization, it introduced a slight size regression in the
generated builds.

Let's make WASM builds use the `rlsf` allocator now, which has been
found to be smaller than `wee_alloc` by two different benchmarks, is
based on an academic paper, and has a much better maintenance status.
This alone reduces the size of the produced WASM modules by 5 KiB, when
compiling with `-Oz` + `wasm-opt` with also `-Oz`.